cleanup shared components
