+++ apps/dotcom/client/src/tla/components/TlaEditor/components/TlaEditorErrorFallback.tsx
@@ -1,0 +1,6 @@
+import { Editor } from 'tldraw'
+
+export function TlaEditorErrorFallback({ error }: { error: unknown; editor?: Editor }) {
+	throw error
+	return null
+}
