+++ apps/dotcom/client/src/tla/components/TlaEditor/components/TlaEditorLegacySharePanel.tsx
@@ -1,0 +1,7 @@
+import { useMaybeApp } from '../../../hooks/useAppState'
+import { TlaEditorTopRightPanel } from '../TlaEditorTopRightPanel'
+
+export function TlaEditorLegacySharePanel() {
+	const app = useMaybeApp()
+	return <TlaEditorTopRightPanel isAnonUser={!app} context="legacy" />
+}
