+++ apps/dotcom/client/src/tla/components/TlaEditor/components/TlaEditorMenuPanel.tsx
@@ -1,0 +1,7 @@
+import { useMaybeApp } from '../../../hooks/useAppState'
+import { TlaEditorTopLeftPanel } from '../TlaEditorTopLeftPanel'
+
+export function TlaEditorMenuPanel() {
+	const app = useMaybeApp()
+	return <TlaEditorTopLeftPanel isAnonUser={!app} />
+}
