+++ apps/dotcom/client/src/tla/components/TlaEditor/components/TlaEditorSharePanel.tsx
@@ -1,0 +1,9 @@
+import { useParams } from 'react-router-dom'
+import { useMaybeApp } from '../../../hooks/useAppState'
+import { TlaEditorTopRightPanel } from '../TlaEditorTopRightPanel'
+
+export function TlaEditorSharePanel() {
+	const app = useMaybeApp()
+	const fileSlug = useParams<{ fileSlug: string }>().fileSlug
+	return <TlaEditorTopRightPanel isAnonUser={!app} context={fileSlug ? 'file' : 'scratch'} />
+}
