+++ apps/dotcom/client/src/tla/components/TlaEditor/editor.module.css
@@ -38,22 +38,3 @@
 .anonStylePanel > button:nth-of-type(1) {
 	margin-top: 4px;
 }
-
-.offlineIndicatorWrapper {
-	position: absolute;
-	top: 0px;
-	left: 0px;
-	width: 100%;
-	padding-top: 3px;
-	height: fit-content;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	pointer-events: none;
-}
-
-@media (max-width: 1080px) {
-	.offlineIndicatorWrapper {
-		padding-top: 44px;
-	}
-}
