+++ apps/dotcom/client/src/tla/components/TlaEditor/top.module.css
@@ -236,3 +236,22 @@ button.nameWidthSetter {
 	pointer-events: none;
 	visibility: hidden;
 }
+
+.offlineIndicatorWrapper {
+	position: absolute;
+	top: 0px;
+	left: 0px;
+	width: 100%;
+	padding-top: 3px;
+	height: fit-content;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	pointer-events: none;
+}
+
+@media (max-width: 1080px) {
+	.offlineIndicatorWrapper {
+		padding-top: 44px;
+	}
+}
