+++ apps/dotcom/client/src/tla/components/TlaFileShareMenu/TlaFileShareMenu.tsx
@@ -25,7 +25,7 @@ export function TlaFileShareMenu({
 	children,
 }: {
 	// this share menu is shown when viewing a file, or a published file (snapshot), or when the logged out user is on the root (the scratchpad)
-	context: 'file' | 'published-file' | 'scratch'
+	context: 'file' | 'published-file' | 'scratch' | 'legacy'
 	fileId?: string
 	source: string
 	children: ReactNode
@@ -55,7 +55,8 @@ export function TlaFileShareMenu({
 
 	const okTabs = {
 		// If the context is a guest file or published file, show the anon share file
-		'anon-share': !isOwner && (context === 'file' || context === 'published-file'),
+		'anon-share':
+			!isOwner && (context === 'file' || context === 'published-file' || context === 'legacy'),
 		export: true,
 		// Can the current user configure the file's sharing settings?
 		share: context === 'file' && fileId && file && isOwner,
